add a rough Julia implementation
This seems to get to be at best around 2/3 as fast as C/OpenMP on a dual
Opteron 6212 + Ubuntu 14.04 machine, but that's likely mostly because I
don't know what I'm doing with Julia.

I suspect the bottleneck is generating the random values, and that a
Julia pro could improve on this quite a bit.